+++ changelog.md
@@ -5,7 +5,7 @@
 - Increase spawned Pokemon duration 15 min -> 30 min
 - Fix some performance issues
 - Fix `/help` bug
-- Add vote, events to tutorial?
+- Add vote, events to tutorial
 - Fix a weird voting bug
 
 TODO: